# Improve performance of compact mode toggling