Instead of showing/hiding [total number of nonmodified tags] one by one, we can toggle a class per OSM element/OSM element version which automatically shows/hides these tags. Thus reducing DOM operations from [total number of nonmodified tags] to [total number of modified elements]

The style change is still slow on /node/.../history but that might be a Firefox issue

Links for testing:
https://www.openstreetmap.org/node/107775/history (300ms => 20ms)
https://www.openstreetmap.org/changeset/165370426 (1300ms+ => <5ms)