@deevroman 2025-04-25
Cool, for /changeset this gives a noticeable speedup. 

For /node, noticeable time is spent on... accessibility:https://share.firefox.dev/4lMNGBe