@deevroman 2025-04-25
fyi: you can move the profiler button to the toolbox so you don't have to open the consolehttps://profiler.firefox.com