@deevroman 2025-04-25
I implemented a workaround: first, the first 8 versions are processed, then I let the browser render them, change the radio button, and then process the remaining ones.